requirements.txt
PySide6==6.9.1
firebase-admin==6.9.0
pyserial==3.5
